+++ src/main/java/roomescape/infrastructure/ThemeJdbcRepository.java
@@ -1,4 +1,4 @@
-package roomescape.dao;
+package roomescape.infrastructure;
 
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -6,6 +6,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.Theme;
+import roomescape.domain.repostiory.ThemeRepository;
 
 import java.util.List;
 import java.util.Map;
